requirements.txt
aiogram
python-dotenv
psycopg2-binary
sqlalchemy
pyyaml